MM-24973 Prevent search input to take full screen on Android landscape
